refactor(docker-jans): remove openbanking distribution code
