+++ docker-jans-auth-server/scripts/bootstrap.py
@@ -2,8 +2,6 @@ import base64
 import json
 import os
 import re
-from urllib.parse import urlparse
-from contextlib import suppress
 from string import Template
 
 from jans.pycloudlib import get_manager
@@ -18,11 +16,7 @@ from jans.pycloudlib.persistence import render_sql_properties
 from jans.pycloudlib.persistence import render_spanner_properties
 from jans.pycloudlib.persistence.utils import PersistenceMapper
 from jans.pycloudlib.utils import cert_to_truststore
-from jans.pycloudlib.utils import get_server_certificate
-from jans.pycloudlib.utils import generate_keystore
 from jans.pycloudlib.utils import as_boolean
-
-from keystore_mod import modify_keystore_path
 
 import logging.config
 from settings import LOGGING_CONFIG
@@ -96,7 +90,6 @@ def main():
             "/etc/jans/conf/jans-couchbase.properties",
         )
         # need to resolve whether we're using default or user-defined couchbase cert
-        # sync_couchbase_cert(manager)
         sync_couchbase_truststore(manager)
 
     if "sql" in persistence_groups:
@@ -125,116 +118,17 @@ def main():
         "changeit",
     )
 
-    # if not os.path.isfile("/etc/certs/idp-signing.crt"):
-    #     manager.secret.to_file("idp3SigningCertificateText", "/etc/certs/idp-signing.crt")
-
-    # manager.secret.to_file("passport_rp_jks_base64", "/etc/certs/passport-rp.jks",
-    #                        decode=True, binary_mode=True)
-
     modify_jetty_xml()
     modify_webdefault_xml()
     configure_logging()
-
-    ext_jwks_uri = os.environ.get("CN_OB_EXT_SIGNING_JWKS_URI", "")
-
-    if ext_jwks_uri:
-        # Open Banking external signing cert and key. Use for generating the PKCS12 and jks keystore
-        ext_cert = "/etc/certs/ob-ext-signing.crt"
-        ext_key = "/etc/certs/ob-ext-signing.key"
-        ext_key_pin = "/etc/certs/ob-ext-signing.pin"
-
-        # Open Banking transport signing cert and key. Use for generating the PKCS12 file.
-        ob_transport_cert = "/etc/certs/ob-transport.crt"
-        ob_transport_key = "/etc/certs/ob-transport.key"
-        ob_transport_pin = "/etc/certs/ob-transport.pin"
-        ob_transport_alias = os.environ.get("CN_OB_AS_TRANSPORT_ALIAS", "OpenBankingAsTransport")
-
-        ob_ext_alias = os.environ.get("CN_OB_EXT_SIGNING_ALIAS", "OpenBanking")
-
-        parsed_url = urlparse(ext_jwks_uri)
-        # uses hostname instead of netloc as netloc may have host:port format
-        hostname = parsed_url.hostname
-
-        # get port listed in netloc or fallback to port 443
-        port = parsed_url.port or 443
-
-        get_server_certificate(
-            hostname,
-            port,
-            "/etc/certs/obextjwksuri.crt"
-        )
-
-        cert_to_truststore(
-            "OpenBankingJwksUri",
-            "/etc/certs/obextjwksuri.crt",
-            "/usr/java/latest/jre/lib/security/cacerts",
-            "changeit",
-        )
-
-        cert_to_truststore(
-            ob_ext_alias,
-            ext_cert,
-            "/usr/java/latest/jre/lib/security/cacerts",
-            "changeit",
-        )
-
-        ext_key_passphrase = ""
-        with suppress(FileNotFoundError):
-            with open(ext_key_pin) as f:
-                ext_key_passphrase = f.read().strip()
-
-        generate_keystore(
-            "ob-ext-signing",
-            manager.config.get("hostname"),
-            manager.secret.get("auth_openid_jks_pass"),
-            jks_fn="/etc/certs/ob-ext-signing.jks",
-            in_key=ext_key,
-            in_cert=ext_cert,
-            alias=ob_ext_alias,
-            in_passwd=ext_key_passphrase,
-        )
-
-        if os.path.isfile(ob_transport_cert):
-            cert_to_truststore(
-                ob_transport_alias,
-                ob_transport_cert,
-                "/usr/java/latest/jre/lib/security/cacerts",
-                "changeit",
-            )
-
-            ob_transport_passphrase = ""
-            with suppress(FileNotFoundError):
-                with open(ob_transport_pin) as f:
-                    ob_transport_passphrase = f.read().strip()
-
-            generate_keystore(
-                "ob-transport",
-                manager.config.get("hostname"),
-                manager.secret.get("auth_openid_jks_pass"),
-                jks_fn="/etc/certs/ob-transport.jks",
-                in_key=ob_transport_key,
-                in_cert=ob_transport_cert,
-                alias=ob_transport_alias,
-                in_passwd=ob_transport_passphrase,
-            )
-
-        keystore_path = "/etc/certs/ob-ext-signing.jks"
-        jwks_uri = ext_jwks_uri
-    else:
-        manager.secret.to_file(
-            "auth_jks_base64",
-            "/etc/certs/auth-keys.jks",
-            decode=True,
-            binary_mode=True,
-        )
-        with open("/etc/certs/auth-keys.json", "w") as f:
-            f.write(base64.b64decode(manager.secret.get("auth_openid_key_base64")).decode())
-
-        keystore_path = "/etc/certs/auth-keys.jks"
-        jwks_uri = f"https://{manager.config.get('hostname')}/jans-auth/restv1/jwks"
-
-    # ensure we're using correct JKS file and JWKS uri
-    modify_keystore_path(manager, keystore_path, jwks_uri)
+    manager.secret.to_file(
+        "auth_jks_base64",
+        "/etc/certs/auth-keys.jks",
+        decode=True,
+        binary_mode=True,
+    )
+    with open("/etc/certs/auth-keys.json", "w") as f:
+        f.write(base64.b64decode(manager.secret.get("auth_openid_key_base64")).decode())
 
     try:
         manager.secret.to_file(
